+++ custom_components/houseplan/const.py
@@ -13,7 +13,7 @@ FILES_URL = "/houseplan_files/files"
 CONTENT_URL = "/api/houseplan/content"
 FILES_DIR = "houseplan/files"
 CONF_ADMIN_ONLY = "admin_only"
-VERSION = "1.43.3"
+VERSION = "1.44.0"
 
 DEFAULT_CONFIG: dict = {
     "spaces": [],